Moved checkable flag handling to the publisher_tree_model where I could remove the column hardcoding.
