+++ rqt_publisher/src/rqt_publisher/publisher_tree_model.py
@@ -32,7 +32,7 @@
 import threading
 
 from python_qt_binding.QtCore import Qt, Signal
-from python_qt_binding.QtGui import QStandardItem
+from python_qt_binding.QtGui import QStandardItem, QStandardItemModel
 
 from rqt_py_common.message_tree_model import MessageTreeModel
 from rqt_py_common.data_items import ReadonlyItem, CheckableItem
@@ -126,3 +126,31 @@ class PublisherTreeModel(MessageTreeModel):
             expression_text = expressions.get(slot_path, repr(slot))
             row[self._column_index['expression']].setText(expression_text)
         return row
+
+    def flags(self, index):
+        flags = QStandardItemModel.flags(self, index) 
+        if (index.column() == self._column_index['expression'] and 
+            index.model().data(index.model().index(index.row(), self._column_index['type'], index.parent()), Qt.DisplayRole) == "bool"):
+            flags = flags | Qt.ItemIsUserCheckable | Qt.ItemIsEditable | Qt.ItemIsEnabled
+        return flags
+    
+    def data(self, index, role):
+        if (index.column() == self._column_index['expression'] and 
+            index.model().data(index.model().index(index.row(), self._column_index['type'], index.parent()), Qt.DisplayRole) == "bool"): 
+            if (role == Qt.CheckStateRole):
+                value = index.model().data(index.model().index(index.row(), index.column(), index.parent()), Qt.DisplayRole)
+                if (value == "True"):
+                    return Qt.Checked
+                else:
+                    return Qt.Unchecked
+        return QStandardItemModel.data(self, index, role)
+    
+    def setData(self, index, value, role):
+        if (index.column() == index.column() == self._column_index['expression'] and 
+            index.model().data(index.model().index(index.row(), self._column_index['type'], index.parent()), Qt.DisplayRole) == "bool"): 
+            if (role == Qt.CheckStateRole):
+                stringVal = "False"
+                if (value == Qt.Checked):
+                    stringVal = "True"
+                return QStandardItemModel.setData(self, index, stringVal, Qt.EditRole)
+        return QStandardItemModel.setData(self, index, value, role)
